+++ app/classes/Propel/Models/Base/Field.php
@@ -1211,9 +1211,15 @@ abstract class Field implements ActiveRecordInterface
             $deleteQuery = ChildFieldQuery::create()
                 ->filterByPrimaryKey($this->getPrimaryKey());
             $ret = $this->preDelete($con);
+            // Fenric\Propel\Behaviors\Eventable behavior
+            if (! fenric('event::model.field.pre.delete')->run([$this, \Propel\Runtime\Propel::getServiceContainer()->getWriteConnection(FieldTableMap::DATABASE_NAME)])) {
+                return 0;
+            }
             if ($ret) {
                 $deleteQuery->delete($con);
                 $this->postDelete($con);
+                // Fenric\Propel\Behaviors\Eventable behavior
+                fenric('event::model.field.post.delete')->run([$this, \Propel\Runtime\Propel::getServiceContainer()->getWriteConnection(FieldTableMap::DATABASE_NAME)]);
                 $this->setDeleted(true);
             }
         });
@@ -1249,6 +1255,10 @@ abstract class Field implements ActiveRecordInterface
         return $con->transaction(function () use ($con) {
             $ret = $this->preSave($con);
             $isInsert = $this->isNew();
+            // Fenric\Propel\Behaviors\Eventable behavior
+            if (! fenric('event::model.field.pre.save')->run([$this, \Propel\Runtime\Propel::getServiceContainer()->getWriteConnection(FieldTableMap::DATABASE_NAME)])) {
+                return 0;
+            }
             if ($isInsert) {
                 $ret = $ret && $this->preInsert($con);
                 // Fenric\Propel\Behaviors\Authorable behavior
@@ -1271,6 +1281,10 @@ abstract class Field implements ActiveRecordInterface
                     }	if (! $this->isColumnModified(FieldTableMap::COL_UPDATED_AT)) {
                         $this->setUpdatedAt(new \DateTime('now'));
                     }
+                // Fenric\Propel\Behaviors\Eventable behavior
+                if (! fenric('event::model.field.pre.insert')->run([$this, \Propel\Runtime\Propel::getServiceContainer()->getWriteConnection(FieldTableMap::DATABASE_NAME)])) {
+                    return 0;
+                }
             } else {
                 $ret = $ret && $this->preUpdate($con);
                 // Fenric\Propel\Behaviors\Authorable behavior
@@ -1285,15 +1299,25 @@ abstract class Field implements ActiveRecordInterface
                     if (! $this->isColumnModified(FieldTableMap::COL_UPDATED_AT)) {
                         $this->setUpdatedAt(new \DateTime('now'));
                     }
+                // Fenric\Propel\Behaviors\Eventable behavior
+                if (! fenric('event::model.field.pre.update')->run([$this, \Propel\Runtime\Propel::getServiceContainer()->getWriteConnection(FieldTableMap::DATABASE_NAME)])) {
+                    return 0;
+                }
             }
             if ($ret) {
                 $affectedRows = $this->doSave($con);
                 if ($isInsert) {
                     $this->postInsert($con);
+                    // Fenric\Propel\Behaviors\Eventable behavior
+                    fenric('event::model.field.post.insert')->run([$this, \Propel\Runtime\Propel::getServiceContainer()->getWriteConnection(FieldTableMap::DATABASE_NAME)]);
                 } else {
                     $this->postUpdate($con);
+                    // Fenric\Propel\Behaviors\Eventable behavior
+                    fenric('event::model.field.post.update')->run([$this, \Propel\Runtime\Propel::getServiceContainer()->getWriteConnection(FieldTableMap::DATABASE_NAME)]);
                 }
                 $this->postSave($con);
+                // Fenric\Propel\Behaviors\Eventable behavior
+                fenric('event::model.field.post.save')->run([$this, \Propel\Runtime\Propel::getServiceContainer()->getWriteConnection(FieldTableMap::DATABASE_NAME)]);
                 FieldTableMap::addInstanceToPool($this);
             } else {
                 $affectedRows = 0;
@@ -2540,6 +2564,8 @@ abstract class Field implements ActiveRecordInterface
      */
     static public function loadValidatorMetadata(ClassMetadata $metadata)
     {
+        fenric('event::model.field.validate')->run([func_get_arg(0), \Propel\Runtime\Propel::getServiceContainer()->getWriteConnection(FieldTableMap::DATABASE_NAME)]);
+
         $metadata->addPropertyConstraint('type', new NotBlank());
         $metadata->addPropertyConstraint('type', new Length(array ('max' => 255,)));
         $metadata->addPropertyConstraint('type', new Regex(array ('pattern' => '/^(?:flag|number|string|text|html|year|date|datetime|time|ip|url|email|image)$/',)));
