+++ app/classes/Propel/Models/Base/Parameter.php
@@ -572,9 +572,15 @@ abstract class Parameter implements ActiveRecordInterface
             $deleteQuery = ChildParameterQuery::create()
                 ->filterByPrimaryKey($this->getPrimaryKey());
             $ret = $this->preDelete($con);
+            // Fenric\Propel\Behaviors\Eventable behavior
+            if (! fenric('event::model.parameter.pre.delete')->run([$this, \Propel\Runtime\Propel::getServiceContainer()->getWriteConnection(ParameterTableMap::DATABASE_NAME)])) {
+                return 0;
+            }
             if ($ret) {
                 $deleteQuery->delete($con);
                 $this->postDelete($con);
+                // Fenric\Propel\Behaviors\Eventable behavior
+                fenric('event::model.parameter.post.delete')->run([$this, \Propel\Runtime\Propel::getServiceContainer()->getWriteConnection(ParameterTableMap::DATABASE_NAME)]);
                 $this->setDeleted(true);
             }
         });
@@ -610,19 +616,37 @@ abstract class Parameter implements ActiveRecordInterface
         return $con->transaction(function () use ($con) {
             $ret = $this->preSave($con);
             $isInsert = $this->isNew();
+            // Fenric\Propel\Behaviors\Eventable behavior
+            if (! fenric('event::model.parameter.pre.save')->run([$this, \Propel\Runtime\Propel::getServiceContainer()->getWriteConnection(ParameterTableMap::DATABASE_NAME)])) {
+                return 0;
+            }
             if ($isInsert) {
                 $ret = $ret && $this->preInsert($con);
+                // Fenric\Propel\Behaviors\Eventable behavior
+                if (! fenric('event::model.parameter.pre.insert')->run([$this, \Propel\Runtime\Propel::getServiceContainer()->getWriteConnection(ParameterTableMap::DATABASE_NAME)])) {
+                    return 0;
+                }
             } else {
                 $ret = $ret && $this->preUpdate($con);
+                // Fenric\Propel\Behaviors\Eventable behavior
+                if (! fenric('event::model.parameter.pre.update')->run([$this, \Propel\Runtime\Propel::getServiceContainer()->getWriteConnection(ParameterTableMap::DATABASE_NAME)])) {
+                    return 0;
+                }
             }
             if ($ret) {
                 $affectedRows = $this->doSave($con);
                 if ($isInsert) {
                     $this->postInsert($con);
+                    // Fenric\Propel\Behaviors\Eventable behavior
+                    fenric('event::model.parameter.post.insert')->run([$this, \Propel\Runtime\Propel::getServiceContainer()->getWriteConnection(ParameterTableMap::DATABASE_NAME)]);
                 } else {
                     $this->postUpdate($con);
+                    // Fenric\Propel\Behaviors\Eventable behavior
+                    fenric('event::model.parameter.post.update')->run([$this, \Propel\Runtime\Propel::getServiceContainer()->getWriteConnection(ParameterTableMap::DATABASE_NAME)]);
                 }
                 $this->postSave($con);
+                // Fenric\Propel\Behaviors\Eventable behavior
+                fenric('event::model.parameter.post.save')->run([$this, \Propel\Runtime\Propel::getServiceContainer()->getWriteConnection(ParameterTableMap::DATABASE_NAME)]);
                 ParameterTableMap::addInstanceToPool($this);
             } else {
                 $affectedRows = 0;
@@ -1119,6 +1143,8 @@ abstract class Parameter implements ActiveRecordInterface
      */
     static public function loadValidatorMetadata(ClassMetadata $metadata)
     {
+        fenric('event::model.parameter.validate')->run([func_get_arg(0), \Propel\Runtime\Propel::getServiceContainer()->getWriteConnection(ParameterTableMap::DATABASE_NAME)]);
+
         $metadata->addPropertyConstraint('code', new NotBlank());
         $metadata->addPropertyConstraint('code', new Length(array ('max' => 255,)));
         $metadata->addPropertyConstraint('code', new Regex(array ('pattern' => '/^[_a-zA-Z0-9-]+$/',)));
