+++ app/classes/Propel/Models/Base/PollVariant.php
@@ -878,9 +878,15 @@ abstract class PollVariant implements ActiveRecordInterface
             $deleteQuery = ChildPollVariantQuery::create()
                 ->filterByPrimaryKey($this->getPrimaryKey());
             $ret = $this->preDelete($con);
+            // Fenric\Propel\Behaviors\Eventable behavior
+            if (! fenric('event::model.poll.variant.pre.delete')->run([$this, \Propel\Runtime\Propel::getServiceContainer()->getWriteConnection(PollVariantTableMap::DATABASE_NAME)])) {
+                return 0;
+            }
             if ($ret) {
                 $deleteQuery->delete($con);
                 $this->postDelete($con);
+                // Fenric\Propel\Behaviors\Eventable behavior
+                fenric('event::model.poll.variant.post.delete')->run([$this, \Propel\Runtime\Propel::getServiceContainer()->getWriteConnection(PollVariantTableMap::DATABASE_NAME)]);
                 $this->setDeleted(true);
             }
         });
@@ -916,6 +922,10 @@ abstract class PollVariant implements ActiveRecordInterface
         return $con->transaction(function () use ($con) {
             $ret = $this->preSave($con);
             $isInsert = $this->isNew();
+            // Fenric\Propel\Behaviors\Eventable behavior
+            if (! fenric('event::model.poll.variant.pre.save')->run([$this, \Propel\Runtime\Propel::getServiceContainer()->getWriteConnection(PollVariantTableMap::DATABASE_NAME)])) {
+                return 0;
+            }
             if ($isInsert) {
                 $ret = $ret && $this->preInsert($con);
                 // Fenric\Propel\Behaviors\Authorable behavior
@@ -938,6 +948,10 @@ abstract class PollVariant implements ActiveRecordInterface
                     }	if (! $this->isColumnModified(PollVariantTableMap::COL_UPDATED_AT)) {
                         $this->setUpdatedAt(new \DateTime('now'));
                     }
+                // Fenric\Propel\Behaviors\Eventable behavior
+                if (! fenric('event::model.poll.variant.pre.insert')->run([$this, \Propel\Runtime\Propel::getServiceContainer()->getWriteConnection(PollVariantTableMap::DATABASE_NAME)])) {
+                    return 0;
+                }
             } else {
                 $ret = $ret && $this->preUpdate($con);
                 // Fenric\Propel\Behaviors\Authorable behavior
@@ -952,15 +966,25 @@ abstract class PollVariant implements ActiveRecordInterface
                     if (! $this->isColumnModified(PollVariantTableMap::COL_UPDATED_AT)) {
                         $this->setUpdatedAt(new \DateTime('now'));
                     }
+                // Fenric\Propel\Behaviors\Eventable behavior
+                if (! fenric('event::model.poll.variant.pre.update')->run([$this, \Propel\Runtime\Propel::getServiceContainer()->getWriteConnection(PollVariantTableMap::DATABASE_NAME)])) {
+                    return 0;
+                }
             }
             if ($ret) {
                 $affectedRows = $this->doSave($con);
                 if ($isInsert) {
                     $this->postInsert($con);
+                    // Fenric\Propel\Behaviors\Eventable behavior
+                    fenric('event::model.poll.variant.post.insert')->run([$this, \Propel\Runtime\Propel::getServiceContainer()->getWriteConnection(PollVariantTableMap::DATABASE_NAME)]);
                 } else {
                     $this->postUpdate($con);
+                    // Fenric\Propel\Behaviors\Eventable behavior
+                    fenric('event::model.poll.variant.post.update')->run([$this, \Propel\Runtime\Propel::getServiceContainer()->getWriteConnection(PollVariantTableMap::DATABASE_NAME)]);
                 }
                 $this->postSave($con);
+                // Fenric\Propel\Behaviors\Eventable behavior
+                fenric('event::model.poll.variant.post.save')->run([$this, \Propel\Runtime\Propel::getServiceContainer()->getWriteConnection(PollVariantTableMap::DATABASE_NAME)]);
                 PollVariantTableMap::addInstanceToPool($this);
             } else {
                 $affectedRows = 0;
@@ -2112,6 +2136,8 @@ abstract class PollVariant implements ActiveRecordInterface
      */
     static public function loadValidatorMetadata(ClassMetadata $metadata)
     {
+        fenric('event::model.poll.variant.validate')->run([func_get_arg(0), \Propel\Runtime\Propel::getServiceContainer()->getWriteConnection(PollVariantTableMap::DATABASE_NAME)]);
+
         $metadata->addPropertyConstraint('title', new NotBlank());
         $metadata->addPropertyConstraint('title', new Length(array ('max' => 255,)));
     }
